Move interface to Contract folder
